requirements.txt
asgiref==3.2.10
astroid==2.4.2
autopep8==1.5.4
colorama==0.4.3
Django==3.1.1
django-cors-headers==3.5.0
djangorestframework==3.12.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
isort==5.5.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
mysql-client==0.0.1
mysql-connector-python==8.0.21
mysqlclient==2.0.1
Pillow==8.0.1
protobuf==3.13.0
pycodestyle==2.6.0
PyJWT==1.7.1
pylint==2.6.0
PyMySQL==0.10.0
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
wrapt==1.12.1
